feat: minimum length check
This represents an optimization that allows for an early exit if the test path is shorter than the minimum length of a path. It also introduces a crate-private Path struct that allows us to cache any calculations that would apply to all routespecs. For now, it is only used to avoid trimming leading and trailing slashes once per routespec. We also cache the number of dots in a routespec.